LayoutTests/platform/ios-wk1/fast/text/system-font-weight-expected.txt

This test makes sure that extra-heavy weights of the system font can be achieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS four.offsetWidth is not three.offsetWidth
PASS five.offsetWidth is not four.offsetWidth
PASS six.offsetWidth is not five.offsetWidth
PASS eight.offsetWidth is not seven.offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
Hello, World
